requirements.txt
django-crispy-forms
crispy-bootstrap4
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap4==2024.10
cssselect==1.3.0
datasets==3.6.0
dill==0.3.8
django==5.0.1
django-bootstrap4==24.1
django-crispy-forms==2.4
django-filter==23.5
evaluate==0.4.3
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hijri-converter==2.3.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.31.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==8.7.0
jinja2==3.1.6
lxml==5.4.0
lxml-html-clean==0.4.2
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mysqlclient==2.2.1
networkx==3.4.2
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parse==1.20.2
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pycparser==2.22
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
qrcode==8.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
rfc3986==1.5.0
safetensors==0.5.3
selenium==4.32.0
sentry-sdk==2.28.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
whitenoise==6.9.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
